// types/job.d.ts
export interface IJobInfo {
  jobName: string;
  salaryDesc: string;
  brandName: string;
  brandStageName: string;
  brandScaleName: string;
  jobLabels: string[];
  skills: string[];
  welfareList: string[];
  bossAvatar: string;
  bossName: string;
  bossTitle: string;
  cityName: string;
  areaDistrict: string;
  businessDistrict: string;
  jobExperience: string;
  jobDegree: string;
  brandLogo: string;
  brandIndustry: string;
}